// src/Types/week.ts
export type Week = {
  id: string
  monday_start?: string
  monday_end?: string
  friday_end?: string
  friday_start?: string
  saturday_end?: string
  saturday_start?: string
  sunday_end?: string
  sunday_start?: string
  thursday_end?: string
  thursday_start?: string
  tuesday_end?: string
  tuesday_start?: string
  wednesday_end?: string
  wednesday_start?: string
}